+++ TinkoffASDKUI/TinkoffASDKUI/BusinessLogic/AddCard/Assembly/AddControllerAssembly.swift
@@ -13,6 +13,7 @@ final class AddCardControllerAssembly: IAddCardControllerAssembly {
 
     private let coreSDK: AcquiringSdk
     private let webFlowControllerAssembly: IThreeDSWebFlowControllerAssembly
+    private let appBasedFlowControllerAssembly: ITDSControllerAssembly
     private let configuration: UISDKConfiguration
 
     // MARK: Init
@@ -20,23 +21,26 @@ final class AddCardControllerAssembly: IAddCardControllerAssembly {
     init(
         coreSDK: AcquiringSdk,
         webFlowControllerAssembly: IThreeDSWebFlowControllerAssembly,
+        appBasedFlowControllerAssembly: ITDSControllerAssembly,
         configuration: UISDKConfiguration
     ) {
         self.coreSDK = coreSDK
         self.webFlowControllerAssembly = webFlowControllerAssembly
+        self.appBasedFlowControllerAssembly = appBasedFlowControllerAssembly
         self.configuration = configuration
     }
 
     // MARK: IAddCardControllerAssembly
 
     func addCardController(customerKey: String) -> IAddCardController {
-        AddCardController(
+        return AddCardController(
             addCardService: coreSDK,
             threeDSDeviceInfoProvider: coreSDK.threeDSDeviceInfoProvider(),
             webFlowController: webFlowControllerAssembly.threeDSWebFlowController(),
             threeDSService: coreSDK,
             customerKey: customerKey,
-            checkType: configuration.addCardCheckType
+            checkType: configuration.addCardCheckType,
+            tdsController: appBasedFlowControllerAssembly.assemble()
         )
     }
 }