+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * [EACQAPW-5415] ChargePaymentProcess Tests
 * [EACQAPW-5418] PaymentSystemResolver Tests
 * [EACQAPW-5429] TinkoffPayController Tests
+* [EACQAPW-4452] App Based Flow Payment support
+* [EACQAPW-4453] Update Third Party bundled xcframeworks
 
 ### Fixed